+++ src/main/java/org/example/remedy/domain/auth/exception/InvalidPasswordException.java
@@ -1,7 +1,7 @@
 package org.example.remedy.domain.auth.exception;
 
-import org.example.remedy.global.config.error.ErrorCode;
-import org.example.remedy.global.config.error.exception.AlreadyExistsException;
+import org.example.remedy.global.error.ErrorCode;
+import org.example.remedy.global.error.exception.AlreadyExistsException;
 
 public class InvalidPasswordException extends AlreadyExistsException {
     public InvalidPasswordException() {
