+++ src/main/java/org/example/remedy/domain/dropping/exception/DroppingNotFoundException.java
@@ -1,7 +1,7 @@
 package org.example.remedy.domain.dropping.exception;
 
-import org.example.remedy.global.config.error.ErrorCode;
-import org.example.remedy.global.config.error.exception.NotFoundException;
+import org.example.remedy.global.error.ErrorCode;
+import org.example.remedy.global.error.exception.NotFoundException;
 
 public class DroppingNotFoundException extends NotFoundException {
     public DroppingNotFoundException() {
