+++ src/main/java/org/example/remedy/domain/user/controller/UserController.java
@@ -1,12 +1,14 @@
 package org.example.remedy.domain.user.controller;
 
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.RequiredArgsConstructor;
 import org.example.remedy.domain.user.dto.request.UserProfileUpdateRequest;
 import org.example.remedy.domain.user.dto.response.UserProfileImageResponse;
 import org.example.remedy.domain.user.dto.response.UserProfileResponse;
 import org.example.remedy.domain.user.service.UserService;
 import org.example.remedy.global.security.auth.AuthDetails;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -36,7 +38,7 @@ public class UserController {
     }
 
     @PutMapping(path = "/profile-image", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
-    public ResponseEntity<UserProfileImageResponse> updateProfileImage(MultipartFile image, @AuthenticationPrincipal AuthDetails authDetails) {
+    public ResponseEntity<UserProfileImageResponse> updateProfileImage(@RequestParam MultipartFile image, @AuthenticationPrincipal AuthDetails authDetails) {
         UserProfileImageResponse res = userService.updateUserProfileImage(image, authDetails.getUser());
         return ResponseEntity.ok(res);
     }
