+++ src/main/java/org/example/remedy/domain/user/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package org.example.remedy.domain.user.exception;
 
-import org.example.remedy.global.config.error.ErrorCode;
-import org.example.remedy.global.config.error.exception.NotFoundException;
+import org.example.remedy.global.error.ErrorCode;
+import org.example.remedy.global.error.exception.NotFoundException;
 
 public class UserNotFoundException extends NotFoundException {
     public UserNotFoundException() {
