+++ src/main/java/org/example/remedy/global/security/jwt/exception/InvalidJwtTokenException.java
@@ -1,7 +1,7 @@
 package org.example.remedy.global.security.jwt.exception;
 
-import org.example.remedy.global.config.error.exception.BusinessBaseException;
-import org.example.remedy.global.config.error.ErrorCode;
+import org.example.remedy.global.error.exception.BusinessBaseException;
+import org.example.remedy.global.error.ErrorCode;
 
 public class InvalidJwtTokenException extends BusinessBaseException {
   public static final BusinessBaseException EXCEPTION =
