+++ src/main/java/org/example/remedy/global/storage/StorageUploader.java
@@ -1,0 +1,7 @@
+package org.example.remedy.global.storage;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface StorageUploader {
+    String upload(MultipartFile file);
+}
